requirements.txt
fastapi
uvicorn
jinja2
pillow
openai
tiktoken
numpy
httpx
